+++ src/component/slider/data.js
